Refactor directory names to match library names
