+++ src/crypto/CryptoInterface.c
@@ -26,7 +26,7 @@
 #include <CryptoInterface.h>
 #include <BPSecLib_Private.h>
 #include <TextUtil.h>
-#include <backend/IdValPair.h>
+#include <dynamic/IdValPair.h>
 
 #include <m-dict.h>
 #include <m-shared-ptr.h>
