+++ src/default_sc/BCB_AES_GCM.c
@@ -32,7 +32,7 @@
 
 #include <BPSecLib_Private.h>
 #include <CryptoInterface.h>
-#include <backend/CBOR.h>
+#include <dynamic/CBOR.h>
 
 #include "DefaultSecContext.h"
 #include "DefaultSecContext_Private.h"
