+++ src/dynamic/AbsSecBlock.c
