+++ src/dynamic/AbsSecBlock.h
