+++ src/dynamic/PublicInterfaceImpl.c
