+++ src/dynamic/PublicInterfaceImpl.h
