+++ src/dynamic/SecOperation.c
