+++ src/dynamic/SecOperation.h
