+++ src/dynamic/TelemetryCounters.c
