+++ src/dynamic/TextUtil.c
