+++ test/DefaultScUtils.h
@@ -25,10 +25,10 @@
 #include <m-string.h>
 
 #include <CryptoInterface.h>
-#include <backend/PublicInterfaceImpl.h>
-#include <backend/SecOperation.h>
-#include <backend/IdValPair.h>
-#include <backend/SecurityActionSet.h>
+#include <dynamic/PublicInterfaceImpl.h>
+#include <dynamic/SecOperation.h>
+#include <dynamic/IdValPair.h>
+#include <dynamic/SecurityActionSet.h>
 #include <mock_bpa/ctr.h>
 
 #include "TestUtils.h"
