+++ test/TestUtils.c
@@ -29,10 +29,10 @@
 #include <TextUtil.h>
 #include <mock_bpa/MockBPA.h>
 
-#include <backend/IdValPair.h>
-#include <backend/SecurityActionSet.h>
-#include <backend/UtilDefs_SeqReadWrite.h>
-#include <policy_provider/SamplePolicyProvider.h>
+#include <dynamic/IdValPair.h>
+#include <dynamic/SecurityActionSet.h>
+#include <dynamic/UtilDefs_SeqReadWrite.h>
+#include <sample_pp/SamplePolicyProvider.h>
 
 #include "TestUtils.h"
