+++ test/test_PublicInterfaceImpl.c
@@ -23,8 +23,8 @@
 
 #include <BPSecLib_Public.h>
 #include <CryptoInterface.h>
-#include <backend/SecurityActionSet.h>
-#include <policy_provider/SamplePolicyProvider.h>
+#include <dynamic/SecurityActionSet.h>
+#include <sample_pp/SamplePolicyProvider.h>
 #include <default_sc/DefaultSecContext.h>
 #include <mock_bpa/agent.h>
 #include <mock_bpa/log.h>
